Refactor imports in message and team controllers to use updated paths; comment out protected routes in message and team routes for future implementation.
