+++ Back-End/Features/Team-Chat/Controllers/message.controller.js
@@ -1,6 +1,6 @@
-import User from "../models/user.model.js"
+import { User } from "../../User-Authentication/models/User.js"
 import Message from "../models/message.model.js";
-import cloudinary from "../lib/cloudinary.js";
+import { v2 as cloudinary } from "cloudinary";
 
 // this route is being protected
 export const getUsersForSidebar = async(req, res)=>{
@@ -42,7 +42,7 @@ export const sendMessages = async(req, res)=>{
       const { text, image } = req.body;
       const { id: receiverId } = req.params;
       const senderId = req.user._id;
-      
+
       let imageUrl;
       if(image){
          // upload base64 image to cloudinary
@@ -66,4 +66,4 @@ export const sendMessages = async(req, res)=>{
       console.log("Erro in sendMessage controller: ", error.message);
       res.status(500).json({message:"Internal Sever Error"});
    }
-} 
+}
