+++ Back-End/Features/Team-Chat/routes/message.route.js
@@ -1,16 +1,16 @@
 import express from "express";
-import {protectRoute} from "../middleware/auth.middleware.js"
+// import {protectRoute} from "../../User-Authentication/middleware/auth.middleware.js";
 import { getMessages, getUsersForSidebar, sendMessages } from "../controllers/message.controller.js";
 
 
 const router = express.Router();
 
 
-router.get("/users", protectRoute, getUsersForSidebar);
+// router.get("/users", protectRoute, getUsersForSidebar);
 // fetches messages with the user id
-router.get("/:id", protectRoute, getMessages);
+// router.get("/:id", protectRoute, getMessages);
 
 // Send message to the user id
-router.post("/send/:id", protectRoute, sendMessages);
+// router.post("/send/:id", protectRoute, sendMessages);
 
 export default router;
